# hexdb_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HexDB REST API",
    "version": "0.1.0",
    "description": "HexDB's REST API. Authenticate with `Authorization: Bearer <API key or session token>` or the `hexdb_session` cookie (admin UI). Errors are `{\"error\": {\"code\", \"message\"}}`. See MANUAL.md for filters, aggregations, permissions and everything else."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:7700"
    }
  ],
...
  "security": [
    {
      "bearer": []
    },
    {
      "session": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key (hxk_...) or a session token (hxs....)."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hexdb_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/changes": {
      "get": {
        "tags": [
          "Changes"
        ],
        "summary": "Committed changes after a sequence number (long poll with wait)",
        "operationId": "get_changes",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tessellation",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "wait",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "An error: {\"error\": {\"code\", \"message\"}}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/changes/stream": {
      "get": {
        "tags": [
          "Changes"
        ],
        "summary": "Committed changes as Server-Sent Events",
        "operationId": "get_changes_stream",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tessellation",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream"
          },
          "default": {
            "description": "An error: {\"error\": {\"code\", \"message\"}}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/streams": {
      "get": {
        "tags": [
          "Streams"
        ],
        "summary": "Streams you can read",
        "operationId": "get_streams",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "An error: {\"error\": {\"code\", \"message\"}}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Streams"
        ],
        "summary": "Create a stream",
        "operationId": "post_streams",
        "requestBody": {
          "required": true,
          "description": "{\"name\", \"description\", \"retention_hours\", \"sources\": [...], \"destinations\": [...]}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "An error: {\"error\": {\"code\", \"message\"}}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}